rename: DiaryTitle 컴포넌트를 DiaryHeader 디렉토리 밑으로 가도록 위치 변경
