+++ src/features/diary/components/content/DiaryEditor/lexical.css
@@ -53,7 +53,7 @@ h1 {
   border-bottom-left-radius: 10px;
   border-bottom-right-radius: 10px;
   /* height: calc(100% - 44px); */
-  height: calc(100vh - 244px);
+  height: calc(100vh - 288px);
   overflow-y: auto;
 }
