+++ src/features/diary/pages/content/DiaryContentPage.tsx
@@ -12,7 +12,8 @@ import { disableScroll, enableScroll } from '@/libs/bodyScrollLock/index.ts';
 import { isSSR } from '@/core/utils/objectUtils.ts';
 import { useMapMarker } from '@/features/diary/hooks/useMapMarker.ts';
 import { useSaveDiary } from '@/features/diary/hooks/useSaveDiary.ts';
-import DiaryTitle from '@/features/diary/components/content/DiaryTitle/DiaryTitle.tsx';
+import DiaryTitle from '@/features/diary/components/content/ContentHeader/DiaryTitle/DiaryTitle.tsx';
+import DiaryDescription from '@/features/diary/components/content/ContentHeader/DiaryDescription/DiaryDescription.tsx';
 
 type Props = {
   documentData: string;
@@ -48,9 +49,8 @@ export default function DiaryContentPage({ diaryId }: Props) {
 
       <MapContext.Provider value={{ map, addMarker, removeMarker, saveDiary }}>
         <div className={style.container}>
-          <div className={style.titleWrapper}>
-            <DiaryTitle metaData={metaData} />
-          </div>
+          <DiaryTitle metaData={metaData} />
+          <DiaryDescription metaData={metaData} />
 
           <section className={style.sectionDivision}>
             <DiaryEditor documentData={documentData || ''} />
